+++ dds/src/dcps/data_representation_builtin_endpoints/mod.rs
@@ -7,7 +7,7 @@ pub mod discovered_reader_data;
 pub mod discovered_topic_data;
 pub mod discovered_writer_data;
 pub mod parameter_id_values;
-pub mod rtps_data;
+pub mod rtps_data_representation;
 pub mod spdp_discovered_participant_data;
 pub mod type_lookup;
